chore(.gitignore): ignore dynamically created nodejs files
NodeJS will create a `node_modules` directory for built dependencies
and these should not be included in the repository as they are
dynamically generated.

The `npm-debug.log` is for local development debugging only so should
not be included in the repository.

Other common files have been ignored like .swp files form Vim and
.DS_Store files from OSX.