feat(i18n): rework loading of locale files for webpack compatability
Instead of having this module responsible for requiring the files on a
given path, it now hands that responsibility off to the file that is
utilising the library.

This means that string paths can be provided on a per project basis
rather than being passed into the module, which caused issues with
webpack loading the resources.

This introduces some breaking changes to the module:

  * `setPath` method is removed as it is now redundant
  * `loadLocales` method has been replaced with `requireLocales` method
    * `requireLocales` accepts an object containing the locales in an
       associative array with the locale name as the key.

The example.js and README.md have been updated to reflect these changes.