tests(i18n): add unit tests to check module functionality
A Jest unit test has been added to ensure the module performs as
expected.

The `i18n.js` file has been updated to throw a correct javascript error
if a locale cannot be found as this is the appropriate response to test.

The README.md has been updated with the instructions of running the
tests and the newly included travis-ci build status.